# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tenants": {
      "get": {
        "operationId": "TenantsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenants"
        ]
      },
      "post": {
        "operationId": "TenantsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tenants"
        ]
      },
      "put": {
        "operationId": "TenantsController_updateAll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenants"
        ]
      }
    },
...
    "/tenants/{id}": {
      "get": {
        "operationId": "TenantsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenants"
        ]
      },
      "put": {
        "operationId": "TenantsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenants"
        ]
      },
      "delete": {
        "operationId": "TenantsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenants"
        ]
      }
    },
    "/tenants/{id}/replace": {
      "put": {
        "operationId": "TenantsController_replaceById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenants"
        ]
      }
    },
    "/tenants/{id}/actors": {
      "get": {
        "operationId": "TenantsController_findActors",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenants"
        ]
      }
    },
    "/tenants/{id}/actor-groups": {
      "get": {
        "operationId": "TenantsController_findActorGroups",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenants"
        ]
      }
    },
    "/tenants/{id}/process-definitions": {
      "get": {
        "operationId": "TenantsController_findProcessDefinitions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenants"
        ]
      }
    },
...
    "/user-task-instances": {
      "get": {
        "operationId": "UserTaskInstancesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserTaskInstances"
        ]
      },
      "post": {
        "operationId": "UserTaskInstancesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserTaskInstanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "UserTaskInstances"
        ]
      }
    },
    "/user-task-instances/{id}": {
      "get": {
        "operationId": "UserTaskInstancesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserTaskInstances"
        ]
      },
      "put": {
        "operationId": "UserTaskInstancesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserTaskInstanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserTaskInstances"
        ]
      },
      "delete": {
        "operationId": "UserTaskInstancesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserTaskInstances"
        ]
      }
    },
    "/user-task-instances/{id}/assignee": {
      "get": {
        "operationId": "UserTaskInstancesController_findAssignee",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserTaskInstances"
        ]
      }
    }
  },
  "info": {
    "title": "API",
    "description": "The API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "CreateUserTaskInstanceDto": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "processInstanceId": {
            "type": "string"
          },
          "assigneeId": {
            "type": "string"
          }
        },
        "required": [
          "taskId",
          "status",
          "tenantId",
          "processInstanceId"
        ]
      },
      "UpdateUserTaskInstanceDto": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "processInstanceId": {
            "type": "string"
          },
          "assigneeId": {
            "type": "string"
          }
        }
      }
    }
  }
}